fixed warnings for [coding]

+++ base/logging.cpp
@@ -86,7 +86,7 @@ namespace my
       s << " " << m_names[level];
 
       double const sec = m_timer.ElapsedSeconds();
-      s << " " << setfill(' ') << setw(16 - m_lens[level]) << sec << " ";
+      s << " " << setfill(' ') << setw(static_cast<int>(16 - m_lens[level])) << sec << " ";
     }
   };
