+++ coding/bzip2_compressor.cpp
@@ -11,8 +11,8 @@ size_t DecompressBZip2IntoFixedSize(char const * pSrc, size_t srcSize, char * pD
 {
   // TODO: Remove unnecessary copying.
   vector<char> src(pSrc, pSrc + srcSize);
-  unsigned int dstUsed = dstSize;
-  int error = BZ2_bzBuffToBuffDecompress(pDst, &dstUsed, &src[0], srcSize, 0, 0);
+  unsigned int dstUsed = static_cast<unsigned int>(dstSize);
+  int error = BZ2_bzBuffToBuffDecompress(pDst, &dstUsed, &src[0], static_cast<unsigned int>(srcSize), 0, 0);
   switch (error)
   {
   case BZ_OK:
@@ -33,10 +33,10 @@ void CompressBZip2(int level, char const * pSrc, size_t srcSize, string & dst)
 {
   ASSERT(level >= 1 && level <= 9, (level));
   dst.resize(srcSize + srcSize / 100 + 699);
-  unsigned int dstUsed = dst.size();
+  unsigned int dstUsed = static_cast<unsigned int>(dst.size());
   // TODO: Remove unnecessary copying.
   vector<char> src(pSrc, pSrc + srcSize);
-  int error = BZ2_bzBuffToBuffCompress(&dst[0], &dstUsed, &src[0], srcSize, level, 0, 0);
+  int error = BZ2_bzBuffToBuffCompress(&dst[0], &dstUsed, &src[0], static_cast<unsigned int>(srcSize), level, 0, 0);
   if (error == BZ_OK)
     dst.resize(dstUsed);
   else
